+++ packages/types/src/index.ts
@@ -7,5 +7,5 @@ export * from './ratings';
 export * from './stats';
 export * from './user';
 export * from './emby';
-export * from './stremio';
+export * from './addon';
 export * from './exclusions';